Add files via upload
Create static table with 5 rows and 5 columns with some initial data
and Write a JavaScript function that accept row, column, (to
identify a particular cell) and a string to update the content of that
cell. If the given row and column is not in the range of 5 x 5, it has
to show error message.